Reorganized code, added license and readme files.
Reorganized api interactions into unqieu module for better
organization in code base. Generated README and license files.
